Updated rand/ package and changed all instnaces of int64 to int in other packages for consistency.

+++ rand/knuth-tests.go
@@ -4,9 +4,9 @@ import (
 	"github.com/phil-mansfield/num"
 )
 
-func frequencyTest(gen *Generator, d, iters int64) float64 {
-	bins := make([]int64, d)
-	for n := int64(0); n < iters; n++ {
+func frequencyTest(gen *Generator, d, iters int) float64 {
+	bins := make([]int, d)
+	for n := 0; n < iters; n++ {
 		bins[gen.UniformInt(0, d - 1)]++
 	}
 
@@ -18,10 +18,10 @@ func frequencyTest(gen *Generator, d, iters int64) float64 {
 	return prob
 }
 
-func FrequencyTest(gen *Generator, d, iters, chiNum int64) []float64 {
+func FrequencyTest(gen *Generator, d, iters, chiNum int) []float64 {
 	chis := make([]float64, chiNum)
 
-	for i := int64(0); i < chiNum; i++ {
+	for i := 0; i < chiNum; i++ {
 		chis[i] = frequencyTest(gen, d, iters / chiNum)
 	}
 
@@ -29,9 +29,9 @@ func FrequencyTest(gen *Generator, d, iters, chiNum int64) []float64 {
 }
 
 
-func serialTest(gen *Generator, d, iters int64) float64 {
-	bins := make([]int64, d * d)
-	for n := int64(0); n < iters; n++ {
+func serialTest(gen *Generator, d, iters int) float64 {
+	bins := make([]int, d * d)
+	for n := 0; n < iters; n++ {
 		x, y := gen.UniformInt(0, d - 1), gen.UniformInt(0, d - 1)
 		bins[y * d + x]++
 	}
@@ -45,10 +45,10 @@ func serialTest(gen *Generator, d, iters int64) float64 {
 	return prob
 }
 
-func SerialTest(gen *Generator, d, iters, chiNum int64) []float64 {
+func SerialTest(gen *Generator, d, iters, chiNum int) []float64 {
 	chis := make([]float64, chiNum)
 
-	for i := int64(0); i < chiNum; i++ {
+	for i := 0; i < chiNum; i++ {
 		chis[i] = serialTest(gen, d, iters / chiNum)
 	}
